\name{events.data}
\alias{events.data}
\docType{data}
\title{activpal events data}
\description{This data frame is an example of the data frame generated when the function activpal.file.reader is used to read in the activpal events file generated by the activpal software
}

\usage{data(events.data)}
\format{
  A data frame with 1001 observations on the following 6 variables.
  \describe{
    \item{\code{time}}{a factor vector}
    \item{\code{datacount}}{a integer vector}
    \item{\code{interval}}{a numeric vector}
    \item{\code{activity}}{a integer vector}
    \item{\code{cumulativesteps}}{a integer vector}
    \item{\code{methrs}}{a numeric vector}
  }
}
\details{Example of file generated by activpal.file.reader function
}

\examples{
	
data("events.data",package="activpalProcessing")

}

\keyword{datasets}
