\name{PhysicalActivityIntervention_8_1}
\alias{PhysicalActivityIntervention_8_1}
\docType{data}

\title{
Example AP Events File}
\description{
This data frame is an example AP events file (as downloaded from the AP software).}

\usage{data(PhysicalActivityIntervention_8_1)}
\format{
  A data frame with 60113 observations on the following 6 variables.
  \describe{
    \item{\code{Time}}{a numeric vector}
    \item{\code{DataCount..samples.}}{a numeric vector}
    \item{\code{Interval..s.}}{a numeric vector}
    \item{\code{ActivityCode..0.sedentary..1..standing..2.stepping.}}{a numeric vector}
    \item{\code{CumulativeStepCount}}{a numeric vector}
    \item{\code{Activity.Score..MET.h.}}{a numeric vector}
  }
}

\examples{
		
data(PhysicalActivityIntervention_8_1)

}
\keyword{datasets}
