\name{second.by.second.data}
\alias{second.by.second.data}
\docType{data}
\title{Second-by-second data frame }
\description{This data frame is an example of the data frame generated when the function second.by.second is applied to an activpal events file
}

\usage{data(second.by.second.data)}
\format{
  A data frame with 696444 observations on the following 9 variables.
  \describe{
    \item{\code{time}}{a factor vector}
    \item{\code{date}}{a factor vector}
    \item{\code{ap.posture}}{a integer vector}
    \item{\code{mets}}{a numeric vector}
    \item{\code{met.hours}}{a numeric vector}
    \item{\code{steps}}{a integer vector}
    \item{\code{fifteen.sec.mets}}{a numeric vector}
    \item{\code{one.min.mets}}{a numeric vector}
    \item{\code{num.events}}{a integer vector}
  }
}
\details{Example of file generated by second.by.second function
}

\examples{
	
#	load data 
data("second.by.second.data",package="activpalProcessing")

}

\keyword{datasets}
